+++ core/agenda-manger/src/main/java/pseudoankit/droid/agendamanger/domain/repository/ReminderRepository.kt
@@ -6,7 +6,7 @@ import java.time.LocalDate
 
 interface ReminderRepository {
     fun getReminders(date: LocalDate): Flow<List<AgendaItem.Reminder>>
-    fun getReminder(id: Int): AgendaItem.Reminder
+    fun getReminder(id: Long): AgendaItem.Reminder
     suspend fun save(payload: AgendaItem.Reminder)
     suspend fun update(payload: AgendaItem.Reminder)
     suspend fun delete(payload: AgendaItem.Reminder)
